add recommended code improvements (u32)

+++ task3/sieve.rs
@@ -6,14 +6,14 @@ fn main() {
 }
 
 
-fn sieve(n: i32) -> Vec<i32> {
+fn sieve(n: u32) -> Vec<u32> {
     let mut work = Vec::new();
     // Vector mit allen Zahlen bis n
     for i in 0..n + 1 {
         work.push(i);
     }
 
-    work[1 as usize] = 0;
+    work[1] = 0;
     for iter in 2..n {
         if work[iter as usize] != 0 {
             let mut demo = 2 * iter;
